Refactor: #18 select 폴더 내로 이동, 모달 컴포넌트 분리
